+++ NEWS.md
@@ -1,6 +1,7 @@
 
-# wrapr 1.1.1 2018-01-06
-
+# wrapr 1.1.1 2018-01-20
+ 
+ * Fix some null/blank substitution issues.
  * Minor documentation improvements.
  
 # wrapr 1.1.0 2018-01-03